GUACAMOLE-641: Leverage singleton instance of LocalEnvironment.
