+++ guacamole/src/main/java/org/apache/guacamole/GuacamoleServletContextListener.java
@@ -30,6 +30,7 @@ import java.util.concurrent.atomic.AtomicReference;
 import javax.inject.Inject;
 import javax.servlet.ServletContextEvent;
 import org.apache.guacamole.environment.Environment;
+import org.apache.guacamole.environment.LocalEnvironment;
 import org.apache.guacamole.extension.ExtensionModule;
 import org.apache.guacamole.log.LogModule;
 import org.apache.guacamole.net.auth.AuthenticationProvider;
@@ -125,7 +126,7 @@ public class GuacamoleServletContextListener extends GuiceServletContextListener
     @Override
     public void contextInitialized(ServletContextEvent servletContextEvent) {
 
-        environment = new WebApplicationEnvironment();
+        environment = LocalEnvironment.getInstance();
 
         // Read configuration information from GUACAMOLE_HOME/guacamole.properties
         try {
